+++ R/designer.R
@@ -20,16 +20,22 @@
 #' index
 design <- function(pheno, size_subset, omit = NULL, iterations = 500,
                    name = "SubSet") {
+  stopifnot(is.numeric(size_subset) && is.finite(size_subset))
+  stopifnot(length(dim(pheno)) == 2)
+  stopifnot(is.numeric(iterations) && is.finite(iterations))
+  stopifnot(is.character(name))
   opt <- Inf
 
   # Calculate batches
   size_data <- nrow(pheno)
-
+  if (size_subset >= size_data) {
+    stop("All the data can be done in one batch.", call. = FALSE)
+  }
   batches <- optimum_batches(size_data, size_subset)
-  size_subset_optimum <- optimum_subset(size_data, batches)
 
-  if (!check_sizes(size_data, size_subset_optimum, batches)) {
-    stop("Please provide a higher number of batches or more samples per batch.")
+  if (!valid_sizes(size_data, size_subset, batches)) {
+    stop("Please provide a higher number of batches or more samples per batch.",
+         call. = FALSE)
   }
 
   pheno_o <- omit(pheno, omit)
@@ -47,9 +53,9 @@ design <- function(pheno, size_subset, omit = NULL, iterations = 500,
   # Numbers are evaluated 4 times, and categories only 3
   # check this on evaluate_index
   eval_n <- ifelse(num, 4, 3)
-
+  size_batches <- internal_batches(size_data, size_subset, batches)
   for (x in seq_len(iterations)) {
-    i <- create_subset(size_data, size_subset_optimum, batches, name = name)
+    i <- create_index(size_data, size_batches, batches, name = name)
 
     subsets <- evaluate_index(i, pheno_o)
     # Evaluate the differences between the subsets and the originals
@@ -87,9 +93,14 @@ design <- function(pheno, size_subset, omit = NULL, iterations = 500,
 #' head(index)
 replicates <- function(pheno, size_subset, controls, omit = NULL,
                        iterations = 500){
-  stopifnot(is.numeric(size_subset))
-  stopifnot(is.numeric(controls))
+  stopifnot(is.numeric(size_subset) && is.finite(size_subset))
+  stopifnot(length(dim(pheno)) == 2)
+  stopifnot(is.numeric(iterations) && is.finite(iterations))
 
+  size_data <- nrow(pheno)
+  if (size_subset >= size_data) {
+    stop("All the data can be done in one batch.", call. = FALSE)
+  }
   if (size_subset < controls) {
     stop("The controls are technical controls for the batches.\n\t",
          "They cannot be above the number of samples per batch.", call. = FALSE)